backend/tests/requirements.txt
pytest==8.2.2
requests==2.32.3
moto==5.0.9
bcrypt==4.1.3
pyjwt==2.8.0